add a test audio mp3 for the media player in focus view.
---------
The MIDI, audio(MP3, OGG) and video files of Bernd Krueger are
licensed under the cc-by-sa Germany License. This means, that
you can use and adapt the files, as long as you contribute to
the copyright holder

Name: Bernd Krueger
Source: http://www.piano-midi.de

The distribution or public playback of the files is only allowed
under identical license conditions.

The scores are open source.